+++ python/dgl/graphbolt/sampled_subgraph.py
@@ -4,6 +4,8 @@ from typing import Dict, Tuple, Union
 
 import torch
 
+from .base import etype_str_to_tuple
+
 
 class SampledSubgraph:
     r"""An abstract class for sampled subgraph. In the context of a
@@ -73,3 +75,131 @@ class SampledSubgraph:
                 heterogeneous edge ids.
         """
         return None
+
+    def exclude_edges(
+        self,
+        edges: Union[
+            Dict[str, Tuple[torch.Tensor, torch.Tensor]],
+            Tuple[torch.Tensor, torch.Tensor],
+        ],
+    ):
+        r"""Exclude edges from the sampled subgraph.
+
+        This function can be used with sampled subgraphs, regardless of whether they
+        have compacted row/column nodes or not. If the original subgraph has
+        compacted row or column nodes, the corresponding row or column nodes in the
+        returned subgraph will also be compacted.
+
+        Parameters
+        ----------
+        self : SampledSubgraph
+            The sampled subgraph.
+        edges : Union[Dict[str, Tuple[torch.Tensor, torch.Tensor]],
+                    Tuple[torch.Tensor, torch.Tensor]]
+            Edges to exclude. If sampled subgraph is homogeneous, then `edges`
+            should be a pair of tensors representing the edges to exclude. If
+            sampled subgraph is heterogeneous, then `edges` should be a dictionary
+            of edge types and the corresponding edges to exclude.
+
+        Returns
+        -------
+        Tuple[node_pairs, reverse_column_node_ids
+                reverse_row_node_ids, reverse_edge_ids]
+            The data tuple is employed for constructing an implemented sampled
+            subgraph.
+        """
+        assert isinstance(self.node_pairs, tuple) == isinstance(edges, tuple), (
+            "The sampled subgraph and the edges to exclude should be both "
+            "homogeneous or both heterogeneous."
+        )
+        # Three steps to exclude edges:
+        # 1. Convert the node pairs to the original ids if they are compacted.
+        # 2. Exclude the edges and get the index of the edges to keep.
+        # 3. Slice the subgraph according to the index.
+        if isinstance(self.node_pairs, tuple):
+            reverse_edges = _to_reverse_ids(
+                self.node_pairs,
+                self.reverse_row_node_ids,
+                self.reverse_column_node_ids,
+            )
+            index = _exclude_homo_edges(reverse_edges, edges)
+            return _slice_subgraph(self, index)
+        else:
+            index = {}
+            for etype, pair in self.node_pairs.items():
+                src_type, _, dst_type = etype_str_to_tuple(etype)
+                reverse_row_node_ids = (
+                    None
+                    if self.reverse_row_node_ids is None
+                    else self.reverse_row_node_ids.get(src_type)
+                )
+                reverse_column_node_ids = (
+                    None
+                    if self.reverse_column_node_ids is None
+                    else self.reverse_column_node_ids.get(dst_type)
+                )
+                reverse_edges = _to_reverse_ids(
+                    pair,
+                    reverse_row_node_ids,
+                    reverse_column_node_ids,
+                )
+                index[etype] = _exclude_homo_edges(
+                    reverse_edges, edges.get(etype)
+                )
+            return _slice_subgraph(self, index)
+
+
+def _to_reverse_ids(node_pair, reverse_row_node_ids, reverse_column_node_ids):
+    u, v = node_pair
+    if reverse_row_node_ids is not None:
+        u = reverse_row_node_ids[u]
+    if reverse_column_node_ids is not None:
+        v = reverse_column_node_ids[v]
+    return (u, v)
+
+
+def _relabel_two_arrays(lhs_array, rhs_array):
+    """Relabel two arrays into a consecutive range starting from 0."""
+    concated = torch.cat([lhs_array, rhs_array])
+    _, mapping = torch.unique(concated, return_inverse=True)
+    return mapping[: lhs_array.numel()], mapping[lhs_array.numel() :]
+
+
+def _exclude_homo_edges(edges, edges_to_exclude):
+    """Return the indices of edges that are not in edges_to_exclude."""
+    # 1. Relabel edges.
+    src, src_to_exclude = _relabel_two_arrays(edges[0], edges_to_exclude[0])
+    dst, dst_to_exclude = _relabel_two_arrays(edges[1], edges_to_exclude[1])
+    # 2. Compact the edges to integers.
+    dst_max_range = dst.numel() + dst_to_exclude.numel()
+    val = src * dst_max_range + dst
+    val_to_exclude = src_to_exclude * dst_max_range + dst_to_exclude
+    # 3. Use torch.isin to get the indices of edges to keep.
+    mask = ~torch.isin(val, val_to_exclude)
+    return torch.nonzero(mask, as_tuple=True)[0]
+
+
+def _slice_subgraph(subgraph: SampledSubgraph, index: torch.Tensor):
+    """Slice the subgraph according to the index."""
+
+    def _index_select(obj, index):
+        if obj is None:
+            return None
+        if isinstance(obj, torch.Tensor):
+            return obj[index]
+        if isinstance(obj, tuple):
+            return tuple(_index_select(v, index) for v in obj)
+        # Handle the case when obj is a dictionary.
+        assert isinstance(obj, dict)
+        assert isinstance(index, dict)
+        ret = {}
+        for k, v in obj.items():
+            ret[k] = _index_select(v, index[k])
+        return ret
+
+    return (
+        _index_select(subgraph.node_pairs, index),
+        subgraph.reverse_column_node_ids,
+        subgraph.reverse_row_node_ids,
+        _index_select(subgraph.reverse_edge_ids, index),
+    )
